requirements.txt
discord.py==1.6.0
Flask==1.1.2
PyNaCl==1.4.0
youtube-dl==2021.4.1
